refactor: moved statistics collection moved statistics collection to ServerActionHandler.cs
