+++ SmartHouseController.MQTT/SmartHouseController.MQTT.Broker/SmartHouseController.MQTT.Broker.Server/Program.cs
@@ -11,10 +11,7 @@ using SmartHouseController.MQTT.Broker.Domain.Entities;
 using SmartHouseController.MQTT.Broker.Server.Extension;
 
 
-
-using var db = new ApplicationDbContext();
-
-var topics = new List<string>()
+var deviceTopics = new List<string>()
 {
     "zigbee2mqtt/0x00158d0006d51e9a",
     "zigbee2mqtt/0x04cf8cdf3c79d013",
@@ -25,8 +22,6 @@ Log.Logger = new LoggerConfiguration()
     .MinimumLevel.Debug()
     .WriteTo.Console()
     .CreateLogger();
-
-Log.Information(Convert.ToDateTime("11.05.2022").ToString());
 
 var mqttFactory = new MqttFactory();
 var mqttServerOptions = new MqttServerOptionsBuilder()
@@ -53,38 +48,10 @@ mqttServer
     .WithValidationHandler()
     .WithDisconnectedHandler()
     .WithConnectedHandler()
-    .WithInterceptingPublishHandler(mqttClient)
+    .WithInterceptingPublishHandler(mqttClient, deviceTopics)
     .WithOnMessageLogHandler();
 
 await mqttClient.ConnectAsync(mqttClientOptions);
 await mqttServer.StartAsync();
-
-mqttServer.InterceptingPublishAsync += e =>
-{
-    var topic = e.ApplicationMessage.Topic;
-    if (!topics.Contains(topic)) return Task.CompletedTask;
-    var payload = Encoding.UTF8.GetString(e.ApplicationMessage!.Payload);
-    var date = DateOnly.FromDateTime(DateTime.Now);
-    var record = db.Statistics.FirstOrDefault(r => r.date == date && r.topic == topic);
-    if (record != null)
-        record.payload = payload;
-    else
-        db.Add(new Staistics
-        {
-            date = date,
-            topic = topic,
-            payload = payload
-        });
-
-    var oldRecords = db.Statistics.Where(r =>
-        r.date.AddMonths(1) <= date
-    ).ToList();
-    
-    db.RemoveRange(oldRecords);
-    
-    db.SaveChanges();
-    
-    return Task.CompletedTask;
-};
 
 while (true) { }
